betterAdmin/requirements.txt
celery==5.4.0
Django==4.2.13
django-cors-headers==4.3.1
django-redis==5.4.0
eventlet==0.36.1
redis==5.0.6
pillow==10.3.0
djangorestframework==3.15.1
channels==4.1.0
daphne==4.1.2
django-filter==24.2
djangorestframework-simplejwt==5.3.1
channels-redis==4.2.0
loguru==0.7.2
requests==2.32.3
PyMySQL==1.1.1
mysqlclient==2.2.4
ua-parser==0.18.0
django-db-connection-pool==1.2.5
sentry-sdk==2.7.1
